Move `goTo` to `PCSearcher`. Remove the usage of `cancelChecked` that throws `Exception`
