+++ utils/src/main/scala/org/alephium/ralph/lsp/utils/IsCancelled.scala
@@ -1,0 +1,6 @@
+// Copyright (c) Alephium
+// SPDX-License-Identifier: LGPL-3.0-only
+
+package org.alephium.ralph.lsp.utils
+
+case class IsCancelled(check: () => Boolean)
